+++ user_management/api/views.py
@@ -10,6 +10,7 @@ from rest_framework import generics, renderers, response, status, views
 from rest_framework.authtoken.models import Token
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.permissions import AllowAny, IsAuthenticated
+from rest_framework.throttling import ScopedRateThrottle
 
 from . import serializers, permissions
 
@@ -19,6 +20,8 @@ User = get_user_model()
 
 class GetToken(ObtainAuthToken):
     renderer_classes = (renderers.JSONRenderer, renderers.BrowsableAPIRenderer)
+    throttle_classes = [ScopedRateThrottle]
+    throttle_scope = 'logins'
 
     def delete(self, request, *args, **kwargs):
         try:
@@ -67,6 +70,8 @@ class PasswordResetEmail(generics.GenericAPIView):
     permission_classes = [permissions.IsNotAuthenticated]
     template_name = 'user_management/password_reset_email.html'
     serializer_class = serializers.PasswordResetEmailSerializer
+    throttle_classes = [ScopedRateThrottle]
+    throttle_scope = 'passwords'
 
     def email_context(self, site, user):
         return {